Fix compilation warnings found by gcc from previous commit. https://tracker.physiomeproject.org/show_bug.cgi?id=2896
git-svn-id: https://svn.physiomeproject.org/svn/cmiss/zinc/library/trunk@10786 3c1deb5b-d424-0410-962d-aba41a686d42

+++ core/source/computed_field/computed_field_mesh_operators.cpp
@@ -110,7 +110,7 @@ public:
 				if (values[i] < 1)
 					return CMZN_ERROR_ARGUMENT;
 			}
-			bool change = (this->numbersOfPoints.size() != valuesCount);
+			bool change = (static_cast<int>(this->numbersOfPoints.size()) != valuesCount);
 			this->numbersOfPoints.resize(valuesCount);
 			for (int i = 0; i < valuesCount; ++i)
 			{
